closes #17; переделал MVC контроллеры на валидацию так как в данном случае обработка ошибок не совсем логична

+++ src/main/java/ru/lebedev/bank/domain/checkingAccount/CheckingAccountMvcController.java
@@ -5,19 +5,21 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.WebDataBinder;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 import ru.lebedev.bank.domain.account.AccountService;
-import ru.lebedev.bank.domain.checkingAccount.CheckingAccountService;
 import ru.lebedev.bank.domain.account.dto.AccountCardTransferFormDTO;
 import ru.lebedev.bank.domain.checkingAccount.dto.CheckingAccountDTO;
 import ru.lebedev.bank.domain.card.CardService;
 import ru.lebedev.bank.domain.card.dto.CardDTO;
 import ru.lebedev.bank.domain.client.ClientService;
 import ru.lebedev.bank.domain.client.dto.ClientDTO;
-import ru.lebedev.bank.domain.transaction.dto.TransactionFormDTO;
+import ru.lebedev.bank.domain.checkingAccount.dto.TransactionFormDTO;
 import ru.lebedev.bank.exception.AccountTransferException;
 import ru.lebedev.bank.exception.CloseDefaultAccountException;
+import ru.lebedev.bank.validator.AccountCardTransferFormDTOValidator;
+import ru.lebedev.bank.validator.TransactionFormDTOValidator;
 
 import javax.validation.Valid;
 import java.math.BigDecimal;
@@ -32,6 +34,18 @@ public class CheckingAccountMvcController {
     private final ClientService clientService;
     private final AccountService accountService;
     private final CardService cardService;
+    private final TransactionFormDTOValidator transactionFormDTOValidator;
+    private final AccountCardTransferFormDTOValidator accountCardTransferFormDTOValidator;
+
+    @InitBinder("phoneTransactionForm")
+    public void initPhoneBinder(WebDataBinder dataBinder) {
+        dataBinder.addValidators(transactionFormDTOValidator);
+    }
+
+    @InitBinder("cardTransactionForm")
+    public void iniCardBinder(WebDataBinder dataBinder) {
+        dataBinder.addValidators(accountCardTransferFormDTOValidator);
+    }
 
 
     @GetMapping
@@ -44,13 +58,13 @@ public class CheckingAccountMvcController {
     @GetMapping("/transfer-phone")
     public String showAccountsTransfer(Model model, Principal principal) {
         List<CheckingAccountDTO> accounts = checkingAccountService.findByClientLogin(principal.getName());
-        model.addAttribute("form", new TransactionFormDTO());
+        model.addAttribute("phoneTransactionForm", new TransactionFormDTO());
         model.addAttribute("accounts", accounts);
         return "accounts/account-transfer";
     }
 
     @PostMapping("/transfer-phone")
-    public String accountsTransfer(@ModelAttribute("form") @Valid TransactionFormDTO transactionFormDTO,
+    public String accountsTransfer(@ModelAttribute("phoneTransactionForm") @Valid TransactionFormDTO transactionFormDTO,
                                    BindingResult bindingResult,
                                    Principal principal, Model model) {
 
@@ -60,16 +74,8 @@ public class CheckingAccountMvcController {
             return "accounts/account-transfer";
         }
 
-        try {
-            accountService.transferMoneyByUserPhoneNumber(transactionFormDTO.getAccount().getId(),
-                    transactionFormDTO.getPhoneNumber(), transactionFormDTO.getAmount());
-        } catch (AccountTransferException e) {
-            bindingResult.rejectValue("phoneNumber", "phoneNumber", "Данного номера нет в системе.");
-            model.addAttribute("form", transactionFormDTO);
-            List<CheckingAccountDTO> accounts = checkingAccountService.findByClientLogin(principal.getName());
-            model.addAttribute("accounts", accounts);
-            return "accounts/account-transfer";
-        }
+        accountService.transferMoneyByUserPhoneNumber(transactionFormDTO.getAccount().getId(),
+                transactionFormDTO.getPhoneNumber(), transactionFormDTO.getAmount());
 
         return "redirect:/client";
     }
@@ -79,14 +85,14 @@ public class CheckingAccountMvcController {
     public String showCardTransfer(Model model, Principal principal) {
         List<CheckingAccountDTO> accounts = checkingAccountService.findByClientLogin(principal.getName());
         List<CardDTO> cards = cardService.findByClientLogin(principal.getName());
-        model.addAttribute("form", new AccountCardTransferFormDTO());
+        model.addAttribute("cardTransactionForm", new AccountCardTransferFormDTO());
         model.addAttribute("accounts", accounts);
         model.addAttribute("cards", cards);
         return "accounts/account-transfer-card";
     }
 
     @PostMapping("/transfer-card")
-    public String cardTransfer(@ModelAttribute("form") @Valid AccountCardTransferFormDTO accountCardTransferFormDTO,
+    public String cardTransfer(@ModelAttribute("cardTransactionForm") @Valid AccountCardTransferFormDTO accountCardTransferFormDTO,
                                BindingResult bindingResult,
                                Principal principal, Model model) {
 
@@ -98,18 +104,8 @@ public class CheckingAccountMvcController {
             return "accounts/account-transfer-card";
         }
 
-        try {
-            accountService.transferMoneyByCardNumber(accountCardTransferFormDTO.getAccountDTO().getId(),
-                    accountCardTransferFormDTO.getCardDTO().getCardNumber(), accountCardTransferFormDTO.getAmount());
-        } catch (AccountTransferException e) {
-            bindingResult.rejectValue("amount", "amount", "На счете не достаточно средств.");
-            model.addAttribute("form", accountCardTransferFormDTO);
-            List<CheckingAccountDTO> accounts = checkingAccountService.findByClientLogin(principal.getName());
-            List<CardDTO> cards = cardService.findByClientLogin(principal.getName());
-            model.addAttribute("accounts", accounts);
-            model.addAttribute("cards", cards);
-            return "accounts/account-transfer-card";
-        }
+        accountService.transferMoneyByCardNumber(accountCardTransferFormDTO.getAccountDTO().getId(),
+                accountCardTransferFormDTO.getCardDTO().getCardNumber(), accountCardTransferFormDTO.getAmount());
 
         return "redirect:/client";
     }
@@ -162,17 +158,15 @@ public class CheckingAccountMvcController {
     public String close(@PathVariable Long accountId, Model model, Principal principal) {
 
         CheckingAccountDTO closeAccount = checkingAccountService.findById(accountId).orElseThrow();
+        if (closeAccount.getIsDefault()) {
+            model.addAttribute("error",
+                    "Данный аккаунт нельзя закрыть, это аккаунт по умолчанию.");
+            return "accounts/account-close";
+        }
         List<CheckingAccountDTO> allAccounts = checkingAccountService.findByClientLogin(principal.getName());
         boolean isOwner = allAccounts.stream().anyMatch(a -> a.getId().equals(closeAccount.getId()));
         if (isOwner) {
-            try {
-                checkingAccountService.deleteById(accountId);
-            } catch (CloseDefaultAccountException e) {
-                model.addAttribute("closeAccount", closeAccount);
-                model.addAttribute("error",
-                        "Данный аккаунт нельзя закрыть, это аккаунт по умолчанию.");
-                return "accounts/account-close";
-            }
+            checkingAccountService.deleteById(accountId);
         }
 
         return "redirect:/client";
